refactor: label 컴포넌트 분리
