+++ apps/client/src/features/project/label/generateRandomColor.ts
