+++ apps/client/src/features/project/useSprintsQuery.ts
@@ -4,6 +4,12 @@ import { projectAPI } from '@/features/project/api.ts';
 export const useSprintsQuery = (projectId: number) => {
   return useQuery({
     queryKey: ['sprints', projectId],
-    queryFn: () => projectAPI.getSprints(projectId),
+    queryFn: async () => {
+      const data = await projectAPI.getSprints(projectId);
+      const { sprints } = data.result;
+
+      return sprints;
+    },
+    throwOnError: true,
   });
 };
